feat(CourseContainer): add `useCourseContext`
