+++ client/app/bundles/course/container/CourseContainer.tsx
@@ -1,27 +1,21 @@
 import { ComponentRef, useEffect, useRef, useState } from 'react';
-import {
-  LoaderFunction,
-  Outlet,
-  useLoaderData,
-  useLocation,
-} from 'react-router-dom';
+import { Outlet, useLocation } from 'react-router-dom';
 import { MenuOutlined } from '@mui/icons-material';
 import { IconButton } from '@mui/material';
-import { CourseLayoutData, SidebarItemData } from 'types/course/courses';
+import { CourseLayoutData } from 'types/course/courses';
 
-import CourseAPI from 'api/course';
 import PopupNotifier from 'course/user-notification/PopupNotifier';
 import Footer from 'lib/components/core/layouts/Footer';
 import { DataHandle, useDynamicNest } from 'lib/hooks/router/dynamicNest';
-import { syncSignals } from 'lib/hooks/unread';
 
 import Breadcrumbs from './Breadcrumbs';
+import { loader, useCourseLoader } from './CourseLoader';
 import Sidebar from './Sidebar';
 
 const CourseContainer = (): JSX.Element => {
   const location = useLocation();
 
-  const data = useLoaderData() as CourseLayoutData;
+  const data = useCourseLoader();
 
   const sidebarRef = useRef<ComponentRef<typeof Sidebar>>(null);
   const [sidebarOpen, setSidebarOpen] = useState(false);
@@ -59,7 +53,7 @@ const CourseContainer = (): JSX.Element => {
         </div>
 
         <div className="flex-grow">
-          <Outlet />
+          <Outlet context={data} />
         </div>
 
         <Footer />
@@ -68,37 +62,6 @@ const CourseContainer = (): JSX.Element => {
       <PopupNotifier />
     </main>
   );
-};
-
-const extractUnreadCountsInto = (
-  record: Record<string, number>,
-  data: SidebarItemData[],
-): Record<string, number> =>
-  data.reduce<Record<string, number>>((unreads, item) => {
-    if (item.unread) unreads[item.key] = item.unread;
-    return unreads;
-  }, record);
-
-const extractUnreadCountsFromLayoutData = (
-  data: CourseLayoutData,
-): Record<string, number> => {
-  const unreads: Record<string, number> = {};
-
-  if (data.sidebar) extractUnreadCountsInto(unreads, data.sidebar);
-  if (data.adminSidebar) extractUnreadCountsInto(unreads, data.adminSidebar);
-
-  return unreads;
-};
-
-const loader: LoaderFunction = async ({ params }) => {
-  const id = parseInt(params?.courseId ?? '', 10) || undefined;
-  if (!id) throw new Error(`CourseContainer was loaded with ID: ${id}.`);
-
-  const response = await CourseAPI.courses.fetchLayout(id);
-
-  syncSignals(extractUnreadCountsFromLayoutData(response.data));
-
-  return response.data;
 };
 
 const handle: DataHandle = (match) => {
